Extract PrefixBadge
