+++ src/devtools/client/debugger/src/components/SecondaryPanes/Breakpoints/BreakpointNavigation.js
@@ -6,131 +6,12 @@ import { selectors } from "ui/reducers";
 import { actions } from "ui/actions";
 import { connect } from "devtools/client/debugger/src/utils/connect";
 import React, { useEffect, useState, useRef } from "react";
-import { createPortal } from "react-dom";
 import MaterialIcon from "ui/components/shared/MaterialIcon";
 const { trackEvent } = require("ui/utils/telemetry");
-
-import { useFeature } from "ui/hooks/settings";
+import { PrefixBadgeButton } from "ui/components/PrefixBadge";
 
 import BreakpointTimeline from "./BreakpointTimeline";
 import { PanelStatus } from "./PanelStatus";
-
-const colors = {
-  blue: "#01ACFD",
-  empty: "#F7F7F7",
-  empty: "#F7F7F7",
-  pink: "#FF87DD",
-  yellow: "#FFE870",
-};
-
-function CircleBadge({ color, onSelect }) {
-  return (
-    <div
-      onClick={() => onSelect(color)}
-      style={{
-        backgroundColor: colors[color],
-        borderRadius: "10px",
-        cursor: "pointer",
-        height: "20px",
-        marginRight: "5px",
-        width: "20px",
-      }}
-    />
-  );
-}
-function EmojiBadge({ emoji, onSelect }) {
-  return (
-    <div
-      onClick={() => onSelect(emoji)}
-      className={`img ${emoji}`}
-      style={{
-        backgroundColor: "#F7F7F7",
-        borderRadius: "10px",
-        cursor: "pointer",
-        height: "20px",
-        marginRight: "5px",
-        width: "20px",
-      }}
-    />
-  );
-}
-
-function PrefixBadgePicker({ onSelect, pickerNode }) {
-  const { top, left } = pickerNode ? pickerNode.getBoundingClientRect() : {};
-
-  return createPortal(
-    <div
-      style={{
-        background: "white",
-        borderRadius: "12px",
-        boxShadow: "0px 1px 2px 0px #00000040",
-        display: "flex",
-        padding: "3px 5px",
-        position: "absolute",
-        right: `${left + 28}px`,
-        top: `${top - 3}px`,
-        zIndex: 100,
-      }}
-    >
-      <EmojiBadge onSelect={onSelect} emoji="rocket" />
-      <EmojiBadge onSelect={onSelect} emoji="unicorn" />
-      <EmojiBadge onSelect={onSelect} emoji="target" />
-      <CircleBadge onSelect={onSelect} color="pink" />
-      <CircleBadge onSelect={onSelect} color="yellow" />
-      <CircleBadge onSelect={onSelect} color="blue" />
-      <CircleBadge onSelect={onSelect} color="empty" />
-    </div>,
-    document.body
-  );
-}
-
-function PrefixBadgeButton({ breakpoint, setBreakpointPrefixBadge }) {
-  const [showPrefixBadge, setShowPrefixBadge] = useState(false);
-  const pickerNode = useRef();
-  const { value: enableUnicornConsole } = useFeature("unicornConsole");
-
-  const prefixBadge = breakpoint.options.prefixBadge;
-  const isEmoji = ["unicorn", "rocket", "target"].includes(prefixBadge);
-  const isColor = ["blue", "pink", "yellow"].includes(prefixBadge);
-
-  if (!enableUnicornConsole) {
-    return null;
-  }
-
-  return (
-    <button
-      className={classnames("h-5 w-5 rounded-full p-px pt-0.5", {
-        img: isEmoji,
-        rocket: prefixBadge === "rocket",
-        target: prefixBadge === "target",
-        unicorn: prefixBadge === "unicorn",
-      })}
-      ref={pickerNode}
-      style={{
-        backgroundColor: isColor ? colors[prefixBadge] : "white",
-        border: "2px solid #c5c5c5",
-        borderRadius: "100%",
-        height: "1.25rem",
-        marginRight: "5px",
-        position: "relative",
-        width: "21px",
-      }}
-      onClick={() => {
-        setShowPrefixBadge(!showPrefixBadge);
-      }}
-    >
-      {showPrefixBadge && (
-        <PrefixBadgePicker
-          pickerNode={pickerNode.current}
-          onSelect={badge => {
-            setShowPrefixBadge(false);
-            setBreakpointPrefixBadge(breakpoint, badge);
-          }}
-        />
-      )}
-    </button>
-  );
-}
 
 function BreakpointNavigation({
   executionPoint,
